data_pipeline/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5
numpy>=1.24.0
google-cloud-storage>=2.10.0
Pillow>=10.0.0
python-magic>=0.4.27 